refactor(landing): move landing page to home page
