+++ package.json
@@ -33,6 +33,7 @@
     "gatsby-plugin-google-gtag": "^1.0.9",
     "gatsby-plugin-intl": "^0.3.2",
     "gatsby-plugin-react-helmet": "^3.1.18",
+    "gatsby-plugin-react-svg": "^2.1.2",
     "gatsby-plugin-sharp": "^2.3.10",
     "gatsby-source-contentful": "^2.1.72",
     "gatsby-source-twitter": "^3.0.0",
