+++ src/components/layout.js
@@ -21,8 +21,6 @@ class Template extends React.Component {
   };
 
   render() {
-    console.log(this.props);
-
     const { children } = this.props;
 
     return (
