+++ src/pages/assets.js
@@ -60,8 +60,6 @@ class RootIndex extends React.Component {
   render() {
     const { location, intl } = this.props;
     const { mediaFilter } = this.state;
-    console.log(this.state);
-    console.log(this.assets[mediaFilter.value]);
     var siteTitle = "Kleros";
     if (mediaFilter.value == "") this.categoryChange(this.categories[0]);
